Add sandbox target, remaining input tools, rename to wayhand-mcp
Sandbox target runs a nested sway compositor and drives it through
wlr virtual pointer, virtual keyboard and screencopy, so the user's real
mouse and keyboard stay free. Every tool takes target=sandbox|desktop.
Adds double_click, right_click, drag, scroll, key and type (clipboard
fallback for text outside the US layout), a shared xkbcommon key map,
and step-sequence builders with unit tests. Binary and env vars renamed
from desktop-driver to wayhand-mcp.

+++ src/inject/mod.rs
@@ -14,7 +14,7 @@ pub enum Button {
 
 #[allow(dead_code)]
 pub trait Injector: Send {
-    fn move_abs(&mut self, x: u16, y: u16) -> Result<()>;
+    fn move_abs(&mut self, x: u32, y: u32) -> Result<()>;
     fn button(&mut self, button: Button, pressed: bool) -> Result<()>;
     fn key(&mut self, code: u16, pressed: bool) -> Result<()>;
     fn scroll(&mut self, horizontal: i32, vertical: i32) -> Result<()>;
